pyproject.toml
[project]
name = "chat-squizzer-bot"
version = "0.1.0"
description = "Bot that gives a short summary of messages"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.2.1",
    "aiosqlite>=0.21.0",
    "nltk>=3.9.2",
    "pymorphy3>=2.0.6",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
]

[tool.isort]
skip = [".gitignore", ".dockerignore", ".env", ".venv", "ENV/", "VENV/"]
multi_line_output = 3
include_trailing_comma = true
line_length = 88
skip_gitignore = true
force_grid_wrap = 0
use_parentheses = true
